+++ src/Google/Api/Ads/AdWords/v201506/AdGroupAdService.php
@@ -3397,7 +3397,7 @@ if (!class_exists("Predicate", false)) {
 if (!class_exists("ProductAd", false)) {
   /**
    * Represents a product ad (known as a <a href=
-   * "//support.google.com/adwords/bin/answer.py?answer=2456103">product
+   * "//support.google.com/adwords/answer/2456103">product
    * listing ad</a> in the AdWords user interface). A product ad displays
    * product data (managed using the Google Merchant Center) that is
    * pulled from the Google base product feed specified in the parent campaign's