+++ src/Google/Api/Ads/AdWords/v201506/CampaignService.php
@@ -4357,7 +4357,7 @@ if (!class_exists("TargetCpaBiddingScheme", false)) {
    * average CPA target.
    * 
    * <p>Note that campaigns must meet <a
-   * href="//support.google.com/adwords/bin/answer.py?answer=2471188">specific
+   * href="//support.google.com/adwords/answer/2471188">specific
    * eligibility requirements</a> before they can use the <code>TargetCpaBiddingScheme</code>
    * bidding strategy.
    * <span class="constraint AdxEnabled">This is disabled for AdX.</span>
@@ -4496,7 +4496,7 @@ if (!class_exists("TargetRoasBiddingScheme", false)) {
    * ensuring that every $1.00 of clicks provides $1.50 in conversion value.
    * 
    * <p>Note that campaigns must meet <a
-   * href="//support.google.com/adwords/bin/answer.py?answer=2471188">specific
+   * href="//support.google.com/adwords/answer/2471188">specific
    * eligibility requirements</a> before they can use the <code>TargetRoasBiddingScheme</code>
    * bidding strategy.
    * <span class="constraint AdxEnabled">This is disabled for AdX.</span>